loongarch: hostreq: Bump Binutils and GCC requirement
Binutils 2.38 is the first release supporting LoongArch.  Maybe we
should make it 2.40 (when the stupid "stack-based relocs" are gone) but
with our cross-compile method it should not make a significant
difference.

GCC 12.1 is the first release supporting LoongArch but it suffers a
severe bug [1] and it may miscompile GCC pass 1, so require 12.2 here.

[1]:https://gcc.gnu.org/PR106096