dependencies: Update for Udev
The Udev build dependencies are directly copied from systemd.  Though
they might be unneeded for a Udev-only build, meson will check them
anyway.  And I'm sure Jinja2 is really used.

In systemd dependencies Attr is dropped because Attr is not used
directly, it's just Acl needs Attr.  OpenSSL is added because udevadm
links to libcrypto.

As "S" and "U" are closer, create a "systemd-udev" entity to simplify
the <phrase> choices.

In Util-linux, lsblk and findmnt links to libudev so it depends on
&systemd-udev;.