gcc: Fix libcc1 description
libcpp is the preprocessor library, but it's a static library which is
only used by GCC itself and not installed.

libcc1 is actually a library for GDB to "compile" expressions, so we can
use fancy expressions in commands, like "print sin(x + 2.0)": the
expression sin(x + 2.0) needs to be "compiled" for evaluation.