pkgmgt: Upgrading Linux API headers cannot be dangerous
Per a discussion in the team, we only consider an upgradation dangerous
if it may render the system unusable.  "Causing something not able to
build" is never considered dangerous.  Thus upgrading some headers
cannot be dangerous.

The Glibc portion will need an update too (it can be upgraded safely
with some caution) to ease security updates.  But let's do the easy
change first...