Remove -v from "chown -R tester" commands
It does no good: normally we have -v for chown so once it no longer has
an effect we can know, but in this case these chown commands will never
have no effect.  And a huge amount of output with -v wastes the server
storage and bandwidth (for both the server and the people reading the
build logs).