vim: Set TERM=xterm-256color for test
I spent some time investigating the difference of vim test results from
different editors.  It turns out the value of TERM can affect the test
results in a deterministic way: when TERM=xterm-256color all tests pass,
when TERM=linux one test fails, and when TERM=vt100 20+ tests fail.

As we are redirecting the output to a file, the actual type of the
terminal does not matter and we can just specify a value known to work.