drm-panic: kernel: Enable SimpleDRM
Currently on a LFS system booted via legacy BIOS (intead of UEFI), the
users can make two different configurations:

1. The DRM driver for the GPU is built as a module.  Then before the
   module is loaded (i.e. the root fs is properly mounted) the kernel
   messages are displayed via the VGA console.
2. The DRM driver for the GPU is built into the kernel image.  Then the
   kernel messages are displayed via the DRM-emulated framebuffer
   console since a very early stage (before the root fs is properly
   mounted).

When the system is booted via UEFI, we use SimpleDRM on the EFI
framebuffer for displaying kernel messages, thus it's always the case 2.

Both are not good for diagnostic in case the root fs cannot be mounted
properly (it's very frequent on lfs-support):

- With 1 the VGA console can only fit 80x24 characters, so the relevent
  information is often flushed away.
- With 2 the panic message is often displayed improperly when a DRM
  driver is running.

The "DRM panic" feature introduced since Linux 6.11 (but it'll only
become usable for LFS since Linux 6.12) can fix the case 2, while case
1 just cannot be fixed.  So start to make LFS booted via legacy BIOS
to use method 2 now, then we can enable DRM panic and fix this
long-standing issue.