drm-panic: kernel: Enable DRM_PANIC
Since the previous commit, we've ensured the kernel displaying via DRM
even in the early stage.  Now enable DRM_PANIC so the kernel will
correctly show the last lines of the kernel messages in case a panic
happens.

Tested with an incorrect root= line in 3 configurations:

- A QEMU VM booted via BIOS
- A real machine booted via BIOS
- A real machine booted via UEFI

And in all configurations the panic message is correctly dumped.
In 1024x768 we can get 48 lines (twice of 24 lines in VGA console).