gettext: Fix build issue with system libxml2
The upstream fix is explicitly checking libxml2 version so it's not
practical to be turned into a sed.  But I don't want to upload another
patch just for such a "temporary" issue (as it'll be fixed for 0.23.1
anyway).  This alternative fix relies on:

1. The fact that the return type of xmlCtxtGetLastError and the arg type
   of xmlStructuredErrorFunc are changed in the same commit
   61034116d0a3 ("error: Make more xmlError structs constant").
2. GCC supports typeof(expr) to get the type of expr w/o really
   evaluating it.

So it should have the same effect as the upstream fix.

Tested with both --with-included-libxml2 and --without-included-libxml2
on BLFS.