Revert "systemd: Update the test suite failures list for 257."
This reverts commit 8f493a63774a200a9be7f1543f85b4fa155ed26f.

When I was building LFS, I somehow forgot to mount /dev/sdb7 to
/mnt/lfs on my system. I ended up chrooting into /mnt/lfs, and it was a
directory, not a filesystem.

This happens from time to time, not sure why I did it this time!

The tests that I had marked as failed no longer fail, but at the same
time, now test-namespace does. Let's undo this commit...