gmp: Fix test count.
This issue of mismatched test count for other installations of
architectures that are not part of the normal installation probably
exists in several XML files throughout the book due to how merges were
done in this branch. This was reported by Visceral on Discord. Though I
am unsure if I should make a point to go through another pass and make
each arch section fall in line with the normal sections.

An alternative is killing the tests for the other architectures. This is
something I've done in GLFS to reduce maintainence burden. But the goal
for many users following this book is to simply get a working system
with capabilities that allow it to run 32-bit software, no hassle. Tests
are thus very important.

Another issue is that the amount of tests can and will differ for
different architectures which I'm pretty sure is not actually considered
presently.

I haven't ran tests in the past but it will probably be a good idea to
start doing it, or cut the losses. I'll be thinking on this.