Makefile: Adjust defaults for ML.
This commit makes ml_all the ARCH default, and `default` got renamed to
`none`. However, `default` is still need for profiling or else the
normal sections refuse to render.

Thus, what happens is ARCH gets set by the Makefile, or the user, the
options available are:
- ml_all
- ml_32
- ml_x32
- none

If `none` is set, R_ARCH is set as `default`, otherwise, R_ARCH is set
as ARCH.