Replaced SysVinit with OpenRC.
This is a rough draft of LFS w/ OpenRC without SysVinit. This is nowhere
close for a merge as I need to confirm this works with LFS. Beyond that,
BLFS also must get support, and BLFS will be a battlefield. Until we get
to that point, please do not have this merged.

This branch has been created because the LSB is getting changed, where
SysVinit will not be part of it. However, OpenRC will be. OpenRC also
provides technology projects like GNOME use. SysVinit simply will not
get those technologies. I believe OpenRC will be the way to go, going
forward for an educational alternative to the Systemd version of LFS.

SysVinit can act as an init system under OpenRC, which would introduce
unneeded complexity. That's why this removes SysVinit.

This shall get rendered at
https://linuxfromscratch.org/~zeckma/. MLFS support is not planned until
merge into trunk.

Feedback is appreciated.